refactor(api): neutralize the billing downgrade service filename
Rename services/zero-billing-downgrade.service.ts to
services/billing-downgrade.service.ts with git mv and update the three
import specifiers. No file contents change.

Closes #28209
