+++ turbo/apps/api/src/signals/routes/billing-downgrade.ts
@@ -13,7 +13,7 @@ import { organizationAuthContext$ } from "../auth/auth-context";
 import { authRoute } from "../auth/auth-route";
 import { publicBrand$ } from "../context/hono";
 import { bodyResultOf } from "../context/request";
-import { downgradeSubscription$ } from "../services/zero-billing-downgrade.service";
+import { downgradeSubscription$ } from "../services/billing-downgrade.service";
 import type { RouteEntry } from "../route-entry";
 
 const adminRequired = Object.freeze({
